Rearrange code a bit.
