+++ src/App.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import BlogPage from './components/containers/BlogPage';
+
+const App = () => (
+  <BlogPage />
+)
+
+export default App;
